Refactored test_fields
