+++ tests/test_shortuuid_field.py
@@ -1,5 +1,5 @@
 import six
-from .test_fields import FieldTestCase
+from .utils import FieldTestCase
 
 from .testapp.models import ShortUUIDTestModel_field, ShortUUIDTestModel_pk, ShortUUIDTestAgregateModel, ShortUUIDTestManyToManyModel
